I've added the CPU temperature
Based on JimKnopfIoT's original script, and basically copying his code, I have managed to adapt it to also display the CPU temperature. It is tested on several Raspberry Pi models. I don't know if it would work on other types of devices, but it is very likely, simply by changing the initial path, just like the original script.